Consolidate classes. Add site files
