+++ .github/workflows/ci.yml
@@ -98,11 +98,11 @@ jobs:
 
       - name: Make target directories
         if: github.event_name != 'pull_request' && (startsWith(github.ref, 'refs/tags/v') || github.ref == 'refs/heads/main')
-        run: mkdir -p target shared/target .js/target kpl-tests/target kpl/target kcl-logging-circe/target kcl-tests/target kinesis-client-logging-circe/target shared-circe/target localstack-aws-v2-test-kit/target kpl-logging-circe/target kinesis-client/target localstack-test-kit-common/target .jvm/target kcl/target .native/target kinesis-client-tests/target localstack-aws-v1-test-kit/target localstack-kinesis-client-test-kit/target shared-ciris/target localstack-kcl-test-kit/target localstack-kpl-test-kit/target compat/target project/target
+        run: mkdir -p target shared/target unidocs/target .js/target site/target kpl-tests/target kpl/target kcl-logging-circe/target kcl-tests/target kinesis-client-logging-circe/target shared-circe/target localstack-aws-v2-test-kit/target kpl-logging-circe/target kinesis-client/target localstack-test-kit-common/target .jvm/target kcl/target .native/target kinesis-client-tests/target localstack-aws-v1-test-kit/target localstack-kinesis-client-test-kit/target shared-ciris/target localstack-kcl-test-kit/target localstack-kpl-test-kit/target compat/target project/target
 
       - name: Compress target directories
         if: github.event_name != 'pull_request' && (startsWith(github.ref, 'refs/tags/v') || github.ref == 'refs/heads/main')
-        run: tar cf targets.tar target shared/target .js/target kpl-tests/target kpl/target kcl-logging-circe/target kcl-tests/target kinesis-client-logging-circe/target shared-circe/target localstack-aws-v2-test-kit/target kpl-logging-circe/target kinesis-client/target localstack-test-kit-common/target .jvm/target kcl/target .native/target kinesis-client-tests/target localstack-aws-v1-test-kit/target localstack-kinesis-client-test-kit/target shared-ciris/target localstack-kcl-test-kit/target localstack-kpl-test-kit/target compat/target project/target
+        run: tar cf targets.tar target shared/target unidocs/target .js/target site/target kpl-tests/target kpl/target kcl-logging-circe/target kcl-tests/target kinesis-client-logging-circe/target shared-circe/target localstack-aws-v2-test-kit/target kpl-logging-circe/target kinesis-client/target localstack-test-kit-common/target .jvm/target kcl/target .native/target kinesis-client-tests/target localstack-aws-v1-test-kit/target localstack-kinesis-client-test-kit/target shared-ciris/target localstack-kcl-test-kit/target localstack-kpl-test-kit/target compat/target project/target
 
       - name: Upload target directories
         if: github.event_name != 'pull_request' && (startsWith(github.ref, 'refs/tags/v') || github.ref == 'refs/heads/main')
@@ -198,3 +198,56 @@ jobs:
 
       - name: Publish
         run: sbt '++ ${{ matrix.scala }}' tlRelease
+
+  site:
+    name: Generate Site
+    strategy:
+      matrix:
+        os: [ubuntu-latest]
+        scala: [2.13.10]
+        java: [temurin@17]
+    runs-on: ${{ matrix.os }}
+    steps:
+      - name: Checkout current branch (full)
+        uses: actions/checkout@v3
+        with:
+          fetch-depth: 0
+
+      - name: Download Java (temurin@17)
+        id: download-java-temurin-17
+        if: matrix.java == 'temurin@17'
+        uses: typelevel/download-java@v2
+        with:
+          distribution: temurin
+          java-version: 17
+
+      - name: Setup Java (temurin@17)
+        if: matrix.java == 'temurin@17'
+        uses: actions/setup-java@v3
+        with:
+          distribution: jdkfile
+          java-version: 17
+          jdkFile: ${{ steps.download-java-temurin-17.outputs.jdkFile }}
+
+      - name: Cache sbt
+        uses: actions/cache@v3
+        with:
+          path: |
+            ~/.sbt
+            ~/.ivy2/cache
+            ~/.coursier/cache/v1
+            ~/.cache/coursier/v1
+            ~/AppData/Local/Coursier/Cache/v1
+            ~/Library/Caches/Coursier/v1
+          key: ${{ runner.os }}-sbt-cache-v2-${{ hashFiles('**/*.sbt') }}-${{ hashFiles('project/build.properties') }}
+
+      - name: Generate site
+        run: sbt '++ ${{ matrix.scala }}' docs/tlSite
+
+      - name: Publish site
+        if: github.event_name != 'pull_request' && github.ref == 'refs/heads/main'
+        uses: peaceiris/actions-gh-pages@v3.9.0
+        with:
+          github_token: ${{ secrets.GITHUB_TOKEN }}
+          publish_dir: site/target/docs/site
+          keep_files: true
