--- kcl-tests/src/it/scala/kinesis4cats/kcl/KCLConsumerSpecResources.scala
@@ -1,28 +1,0 @@
-/*
- * Copyright 2023-2023 etspaceman
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package kinesis4cats.kcl
-
-import cats.effect.Deferred
-import cats.effect.std.Queue
-
-import kinesis4cats.client.KinesisClient
-
-final case class KCLConsumerSpecResources[F[_]](
-    client: KinesisClient[F],
-    deferredStarted: Deferred[F, Unit],
-    resultsQueue: Queue[F, CommittableRecord[F]]
-)
