+++ kinesis-client/src/main/scala/kinesis4cats/client/KinesisClient.scala
@@ -48,7 +48,7 @@ class KinesisClient[F[_]] private[kinesis4cats] (
     logger: StructuredLogger[F]
 )(implicit
     F: Async[F],
-    LE: KinesisClientLogEncoders
+    LE: KinesisClient.LogEncoders
 ) {
 
   private def requestLogs[A: LogEncoder](
@@ -318,7 +318,7 @@ object KinesisClient {
     * @param F
     *   F with an [[cats.effect.Async Async]] instance
     * @param LE
-    *   [[kinesis4cats.client.KinesisClientLogEncoders KinesisClientLogEncoders]]
+    *   [[kinesis4cats.client.KinesisClient.LogEncoders LogEncoders]]
     * @return
     *   [[cats.effect.Resource Resource]] containing a
     *   [[kinesis4cats.client.KinesisClient]]
@@ -327,9 +327,144 @@ object KinesisClient {
       client: KinesisAsyncClient
   )(implicit
       F: Async[F],
-      LE: KinesisClientLogEncoders
+      LE: LogEncoders
   ): Resource[F, KinesisClient[F]] = for {
     clientResource <- Resource.fromAutoCloseable(F.pure(client))
     logger <- Slf4jLogger.create[F].toResource
   } yield new KinesisClient[F](clientResource, logger)
+
+  /** Helper class containing required
+    * [[kinesis4cats.logging.LogEncoder LogEncoders]] for the
+    * [[kinesis4cats.client.KinesisClient KinesisClient]]
+    */
+  final class LogEncoders(implicit
+      val addTagsToStreamRequestLogEncoder: LogEncoder[AddTagsToStreamRequest],
+      val addTagsToStreamResponseLogEncoder: LogEncoder[
+        AddTagsToStreamResponse
+      ],
+      val createStreamRequestLogEncoder: LogEncoder[CreateStreamRequest],
+      val createStreamResponseLogEncoder: LogEncoder[CreateStreamResponse],
+      val decreaseStreamRetentionPeriodRequestLogEncoder: LogEncoder[
+        DecreaseStreamRetentionPeriodRequest
+      ],
+      val decreaseStreamRetentionPeriodResponseLogEncoder: LogEncoder[
+        DecreaseStreamRetentionPeriodResponse
+      ],
+      val deleteStreamRequestLogEncoder: LogEncoder[DeleteStreamRequest],
+      val deleteStreamResponseLogEncoder: LogEncoder[DeleteStreamResponse],
+      val deregisterStreamConsumerRequestLogEncoder: LogEncoder[
+        DeregisterStreamConsumerRequest
+      ],
+      val deregisterStreamConsumerResponseLogEncoder: LogEncoder[
+        DeregisterStreamConsumerResponse
+      ],
+      val describeLimitsRequestLogEncoder: LogEncoder[DescribeLimitsRequest],
+      val describeLimitsResponseLogEncoder: LogEncoder[DescribeLimitsResponse],
+      val describeStreamRequestLogEncoder: LogEncoder[DescribeStreamRequest],
+      val describeStreamResponseLogEncoder: LogEncoder[DescribeStreamResponse],
+      val describeStreamConsumerRequestLogEncoder: LogEncoder[
+        DescribeStreamConsumerRequest
+      ],
+      val describeStreamConsumerResponseLogEncoder: LogEncoder[
+        DescribeStreamConsumerResponse
+      ],
+      val describeStreamSummaryRequestLogEncoder: LogEncoder[
+        DescribeStreamSummaryRequest
+      ],
+      val describeStreamSummaryResponseLogEncoder: LogEncoder[
+        DescribeStreamSummaryResponse
+      ],
+      val disableEnhancedMonitoringRequestLogEncoder: LogEncoder[
+        DisableEnhancedMonitoringRequest
+      ],
+      val disableEnhancedMonitoringResponseLogEncoder: LogEncoder[
+        DisableEnhancedMonitoringResponse
+      ],
+      val enableEnhancedMonitoringRequestLogEncoder: LogEncoder[
+        EnableEnhancedMonitoringRequest
+      ],
+      val enableEnhancedMonitoringResponseLogEncoder: LogEncoder[
+        EnableEnhancedMonitoringResponse
+      ],
+      val getRecordsRequestLogEncoder: LogEncoder[GetRecordsRequest],
+      val getRecordsResponseLogEncoder: LogEncoder[GetRecordsResponse],
+      val getShardIteratorRequestLogEncoder: LogEncoder[
+        GetShardIteratorRequest
+      ],
+      val getShardIteratorResponseLogEncoder: LogEncoder[
+        GetShardIteratorResponse
+      ],
+      val increaseStreamRetentionPeriodRequestLogEncoder: LogEncoder[
+        IncreaseStreamRetentionPeriodRequest
+      ],
+      val increaseStreamRetentionPeriodResponseLogEncoder: LogEncoder[
+        IncreaseStreamRetentionPeriodResponse
+      ],
+      val listShardsRequestLogEncoder: LogEncoder[ListShardsRequest],
+      val listShardsResponseLogEncoder: LogEncoder[ListShardsResponse],
+      val listStreamConsumersRequestLogEncoder: LogEncoder[
+        ListStreamConsumersRequest
+      ],
+      val listStreamConsumersResponseLogEncoder: LogEncoder[
+        ListStreamConsumersResponse
+      ],
+      val listStreamsRequestLogEncoder: LogEncoder[ListStreamsRequest],
+      val listStreamsResponseLogEncoder: LogEncoder[ListStreamsResponse],
+      val listTagsForStreamRequestLogEncoder: LogEncoder[
+        ListTagsForStreamRequest
+      ],
+      val listTagsForStreamResponseLogEncoder: LogEncoder[
+        ListTagsForStreamResponse
+      ],
+      val mergeShardsRequestLogEncoder: LogEncoder[MergeShardsRequest],
+      val mergeShardsResponseLogEncoder: LogEncoder[MergeShardsResponse],
+      val putRecordRequestLogEncoder: LogEncoder[PutRecordRequest],
+      val putRecordResponseLogEncoder: LogEncoder[PutRecordResponse],
+      val putRecordsRequestLogEncoder: LogEncoder[PutRecordsRequest],
+      val putRecordsResponseLogEncoder: LogEncoder[PutRecordsResponse],
+      val registerStreamConsumerRequestLogEncoder: LogEncoder[
+        RegisterStreamConsumerRequest
+      ],
+      val registerStreamConsumerResponseLogEncoder: LogEncoder[
+        RegisterStreamConsumerResponse
+      ],
+      val removeTagsFromStreamRequestLogEncoder: LogEncoder[
+        RemoveTagsFromStreamRequest
+      ],
+      val removeTagsFromStreamResponseLogEncoder: LogEncoder[
+        RemoveTagsFromStreamResponse
+      ],
+      val splitShardRequestLogEncoder: LogEncoder[SplitShardRequest],
+      val splitShardResponseLogEncoder: LogEncoder[SplitShardResponse],
+      val startStreamEncryptionRequestLogEncoder: LogEncoder[
+        StartStreamEncryptionRequest
+      ],
+      val startStreamEncryptionResponseLogEncoder: LogEncoder[
+        StartStreamEncryptionResponse
+      ],
+      val stopStreamEncryptionRequestLogEncoder: LogEncoder[
+        StopStreamEncryptionRequest
+      ],
+      val stopStreamEncryptionResponseLogEncoder: LogEncoder[
+        StopStreamEncryptionResponse
+      ],
+      val subscribeToShardRequestLogEncoder: LogEncoder[
+        SubscribeToShardRequest
+      ],
+      val subscribeToShardResponseLogEncoder: LogEncoder[
+        SubscribeToShardResponse
+      ],
+      val updateShardCountRequestLogEncoder: LogEncoder[
+        UpdateShardCountRequest
+      ],
+      val updateShardCountResponseLogEncoder: LogEncoder[
+        UpdateShardCountResponse
+      ],
+      val updateStreamModeRequestLogEncoder: LogEncoder[
+        UpdateStreamModeRequest
+      ],
+      val updateStreamModeResponseLogEncoder: LogEncoder[
+        UpdateStreamModeResponse
+      ]
+  )
 }
