+++ kpl-logging-circe/src/main/scala/kinesis4cats/kpl/logging/instances/circe.scala
@@ -76,6 +76,6 @@ object circe {
     Json.obj(fields.toSeq: _*)
   }
 
-  implicit val kplProducerLogEncoders: KPLProducerLogEncoders =
-    new KPLProducerLogEncoders()
+  implicit val kplProducerLogEncoders: KPLProducer.LogEncoders =
+    new KPLProducer.LogEncoders()
 }
